refactor(test): api 테스트 페이지 폴더 안으로 이동 #313
